% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_data.R
\name{get_data}
\alias{get_data}
\title{Get current and forecasted weather data for a specified city}
\usage{
get_data(city, forecast_horizon = 0, api_key = FALSE)
}
\arguments{
\item{city}{a character of the city for that weather needs to be fetched}

\item{forecast_horizon}{An integer between 0 and 7 to specify the maximum forecast horizon in days.}

\item{api_key}{a character of a valid 'Apixu' API key. If FALSE, the function uses the included standard API key.}
}
\value{
a list containing three main parts: 1) metaData: geographical information of the city,
2) current: current weather information, 3) forecast: forecast weather information
}
\description{
This function requires a character of a city and optionally a forecast horizon to return a list with
geographical information about the city, data about the current weather and if specified also detailed
information for the forecasted time periode.
}
\examples{
get_data("Bangalore", 1)

}
